combine all files
