+++ CMakeLists.txt
@@ -354,10 +354,9 @@ endforeach()
 if(USE_DSE_PLATFORM)
 	project(dse)
 	set(DSE_PLATFORM_DIR ${CMAKE_CURRENT_SOURCE_DIR}/platform/dse)
-	set(DSE_PLATFORM_INCLUDE_DIR ${DSE_PLATFORM_DIR}/include)
-	include_directories(${DSE_PLATFORM_INCLUDE_DIR})
 	set(DSE_PLATFORM_SCRIPT_DIR ${DSE_PLATFORM_DIR}/script)
-	file(GLOB DSE_PLATFORM_SRCS ${DSE_PLATFORM_DIR}/src/*.c)
+	include_directories(${DSE_PLATFORM_DIR})
+	file(GLOB DSE_PLATFORM_SRCS ${DSE_PLATFORM_DIR}/dse.c)
 	add_executable(${PROJECT_NAME} ${DSE_PLATFORM_SRCS})
 	set(DSE_PLATFORM_LIBS event)
 	set(DSE_PLATFORM_LIBS ${DSE_PLATFORM_LIBS} minikonoha)
