--- platform/dse/src/util.c
@@ -1,49 +1,0 @@
-/****************************************************************************
- * Copyright (c) 2012, the Konoha project authors. All rights reserved.
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- *  * Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *  * Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
- * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
- * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- ***************************************************************************/
-
-/* ************************************************************************ */
-
-#include "dse.h"
-
-size_t totalMalloc = 0;
-
-void *dse_malloc(size_t size)
-{
-	void *ptr = malloc(size);
-	if (ptr == NULL) {
-		DEBUG_PRINT("malloc failed");
-		size = 0;
-	}
-	totalMalloc += size;
-	DEBUG_PRINT("totalMalloc:%ld @dse_malloc()", totalMalloc);
-	return ptr;
-}
-
-void dse_free(void *ptr, size_t size)
-{
-	free(ptr);
-	totalMalloc -= size;
-	DEBUG_PRINT("totalMalloc:%ld @dse_free()", totalMalloc);
-	DEBUG_ASSERT(totalMalloc >= 0);
-}
